Use a main file for main loop instead of init
