+++ setup.py
@@ -53,5 +53,5 @@ setup(
     ],
     package_data={'sequali': ['*.c', '*.h', '*.pyi', 'py.typed']},
     ext_modules=[Extension("sequali._qc", ["src/sequali/_qcmodule.c"])],
-    entry_points={"console_scripts": ['sequali=sequali:main']},
+    entry_points={"console_scripts": ['sequali=sequali.__main__:main']},
 )
